total-spending-api-refactor
 - [x] remove unused loc
 - [x] query reduced
 - [x] spelling changed from spendings to spending
